:recycle: Split LoopProcedure
